requirements/base.txt
# Django
django-environ
Pillow
channels[daphne]
channels-redis
django-celery-results
django-celery-beat
django-redis
djangorestframework
django-filter
markdown
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
